requirements.txt
datasets
huggingface_hub
pandas
pillow
requests
tqdm